Simple VGA image format
I designed a custom image format for 256 color VGA images.  There's a python
script to convert BMP files to it (RLI file) and a subroutine in the bootloader
to read and display it.  The format contains the image dimensions, 256 color
palette and run length encoded image data.  The script converts the 8 bit RGB
colors to 6 bit by dividing them with four (right shifting two positions).